improve the test of MetadataController.

+++ paimon-web-server/src/main/java/org/apache/paimon/web/server/data/vo/DataFileVO.java
@@ -32,7 +32,7 @@ public class DataFileVO {
 
     private String partition;
 
-    private long bucket;
+    private Integer bucket;
 
     private String filePath;
